refactor(WinningStatistics): 로또 통계 계산 책임 분리 - 로또 통계 계산 책임을 `printResult`에서  `WinningStatistics`로 이동
