+++ src/main/kotlin/lotto/model/WinningStatistics.kt
@@ -1,0 +1,22 @@
+package lotto.model
+
+class WinningStatistics(
+    lottoBundle: List<Lotto>,
+    winningLotto: WinningLotto,
+) {
+    private val statistics: Map<Rank, Int> = calculateStatistics(lottoBundle, winningLotto)
+
+    private fun calculateStatistics(
+        lottoBundle: List<Lotto>,
+        winningLotto: WinningLotto,
+    ): Map<Rank, Int> {
+        val result = Rank.entries.associateWith { 0 }.toMutableMap()
+        lottoBundle.forEach { lotto ->
+            val rank = Rank.fromMatch(winningLotto.getCountOfMatch(lotto), winningLotto.isBonusNumberMatch(lotto))
+            result[rank] = result[rank]!! + 1
+        }
+        return result
+    }
+
+    fun getAllStatistics(): Map<Rank, Int> = statistics
+}
